requirements.txt
Pillow==9.5.0
qrcode[pil]